MID: 574 MSG: Filtratge del valors a la pestanya "Sense configurar" de la pàgina de configuració d'àrees
git-svn-id: svn://svn.code.sf.net/p/gehelium/code/helium/branches/helium-2.6@942 e3271b2b-fd41-4d56-bedf-4b5636213925

+++ helium-webapp/src/main/java/net/conselldemallorca/helium/webapp/mvc/MonitorController.java
@@ -68,7 +68,7 @@ public class MonitorController extends BaseController {
 			for (int a = 0; a < ids.length; ++a) {
 				hs.add(bean.getThreadCpuTime(ids[a]));
 			}
-			long tiempoCPUTotal = Collections.max(hs);
+			long tiempoCPUTotal =  ((Long)Collections.max(hs)).longValue();
 			for (int a = 0; a < ids.length; ++a) {
 				String nombre = (info[a].getLockName() == null ? info[a].getThreadName() : info[a].getLockName());
 				if (!"main".equals(nombre)) {
